% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{allowed_amino_acids}
\alias{allowed_amino_acids}
\title{List allowed amino acids}
\usage{
allowed_amino_acids()
}
\description{
Returns a data.frame of allowed amino acid single-letter codes. These come
from IUPAC (International Union of Pure and Applied Chemistry),
\href{https://www.bioinformatics.org/sms/iupac.html}{see here} for details.
}
