% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_nohet_variant_string}
\alias{check_nohet_variant_string}
\title{Check that no heterozygous calls within a variant string}
\usage{
check_nohet_variant_string(x)
}
\arguments{
\item{x}{a variant string or vector of variant strings.}
}
\description{
Check that no heterozygous calls within a variant string.
}
