% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_position_string}
\alias{check_position_string}
\title{Check for a valid position string}
\usage{
check_position_string(x)
}
\arguments{
\item{x}{a character string or vector of character strings.}
}
\description{
Checks that an input string (or a vector of strings) matches the required
format for a position string. This is equivalent to a full variant string but
with the amino acid information removed, so just giving the gene name(s) and
position(s).
}
