% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compare_position_string}
\alias{compare_position_string}
\title{Compares a position strings against variant strings to look for a match}
\usage{
compare_position_string(target_string, comparison_strings)
}
\arguments{
\item{target_string}{a single position string that we want to compare.}

\item{comparison_strings}{a vector of variant strings against which the
target is compared.}
}
\description{
Compares a target position string against a vector of comparison strings. A
match is found if every codon position in every gene of the target is also
found within the comparison (irrespective of the observed amino acids).
}
