% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{count_het_loci}
\alias{count_het_loci}
\title{Count number of heterozygous loci}
\usage{
count_het_loci(x)
}
\arguments{
\item{x}{a vector of variant strings.}
}
\description{
Count the number of heterozygous loci in a variant string. This is over all
genes and all codon positions.
}
