% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{count_unphased_hets}
\alias{count_unphased_hets}
\title{Count the number of unphased heterozygous loci}
\usage{
count_unphased_hets(x)
}
\arguments{
\item{x}{a variant string or vector of variant strings. Note, these are not internally checked for being valid variant strings, it is up to the user to ensure this (see \code{?check_variant_string}).}
}
\description{
Count the number of unphased heterozygous loci in each variant string. Return
the number as a vector.
}
