% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{drop_read_counts}
\alias{drop_read_counts}
\title{Drop read counts from a variant string.}
\usage{
drop_read_counts(x)
}
\arguments{
\item{x}{a variant string or vector of variant strings.}
}
\description{
Takes a vector of variant strings and strips and information on read counts.
}
