% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_single_locus_variants}
\alias{extract_single_locus_variants}
\title{Extract all single-locus variants from a variant string}
\usage{
extract_single_locus_variants(x)
}
\arguments{
\item{x}{a vector of variant strings. Note, these are not internally checked for being valid variant strings, it is up to the user to ensure this (see \code{?check_variant_string}).}
}
\description{
Takes a vector of variant strings, potentially with information at multiple
codon positions or genes, and returns variant strings corresponding to all
unique single-locus variants within the input. For example, crt:72_73:C_N/V
can be extracted to crt:72:C, crt:73:N, and crt:73:V.
}
