% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_component_variants}
\alias{get_component_variants}
\title{Get all genotypes that are consistent with a variant string}
\usage{
get_component_variants(x)
}
\arguments{
\item{x}{a vector of variant strings.}
}
\description{
For a variant string with at most one heterozygous locus we can unambiguously
define the genotypes that are present in this mixture. This function returns
all such component genotypes.
}
