% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_consistent_variants}
\alias{get_consistent_variants}
\title{Get all genotypes that are consistent with a variant string}
\usage{
get_consistent_variants(x)
}
\arguments{
\item{x}{a vector of variant strings.}
}
\description{
For a variant string with heterozygous loci there are several genotypes that
may be present in this mixture. This function returns all possible genotypes
that may be found within the variant. If there is a single heterozygous site
then two genotypes will be produced, and we can be 100% confident that both
are present. For 2 or more heterozygous loci we can never be 100% sure of the
combinations present.
}
