% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{long_to_position}
\alias{long_to_position}
\title{Take long form information and convert to position string}
\usage{
long_to_position(x)
}
\arguments{
\item{x}{a list of data frames corresponding to position strings.}
}
\description{
Takes a list of data frames in long form and converts each to position string
format.
}
