% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{long_to_variant}
\alias{long_to_variant}
\title{Take long form information and convert to variant string}
\usage{
long_to_variant(x)
}
\arguments{
\item{x}{a list of data.frames.}
}
\description{
Takes a list of data.frames in long form and converts each to variant string
format.
}
