% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{overlay_variant}
\alias{overlay_variant}
\title{Overlay one variant string onto another}
\usage{
overlay_variant(var1, var2)
}
\arguments{
\item{var1}{the first variant string, which overwrites in the case of shared positions.}

\item{var2}{the second variant string, which is overwritten in the case of shared positions.}
}
\description{
All the positions and amino-acids from the first variant are combined with
the second, overwriting where positions are the same.
}
\details{
Note, inputs are not internally checked for being valid variant strings, it is up to the user to ensure this (see \code{?check_variant_string}).
}
