% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{position_to_long}
\alias{position_to_long}
\title{Expand position strings into long form data.frames}
\usage{
position_to_long(x)
}
\arguments{
\item{x}{a vector of position strings. Note, these are not internally checked for being valid position strings, it is up to the user to ensure this (see \code{?check_position_string}).}
}
\description{
Takes a vector of position strings and expands into a list of data.frames
containing the same information in long form.
}
