% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{subset_position}
\alias{subset_position}
\title{Subset position of a variant string}
\usage{
subset_position(position_string, variant_strings)
}
\arguments{
\item{position_string}{a single position string.}

\item{variant_strings}{a variant string or vector of variant strings.}
}
\description{
Given a vector of variant strings and a single position string, subsets all
variant strings to only the genes and codons in the position string. Retains
read counts at these positions if present.
}
