% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{variant_to_long}
\alias{variant_to_long}
\title{Expand variant strings into long form data.frames}
\usage{
variant_to_long(x)
}
\arguments{
\item{x}{a vector of variant strings. Note, these are not internally checked for being valid variant strings, it is up to the user to ensure this (see \code{?check_variant_string}).}
}
\description{
Takes a vector of variant strings and expands into a list of data.frames
containing the same information in long form.
}
