% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_variant_string}
\alias{check_variant_string}
\title{Check for a valid variant string}
\usage{
check_variant_string()
}
\description{
Checks that an input string matches the required format for a variant string.
}
