% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{count_het_loci}
\alias{count_het_loci}
\title{Count the number of heterozygous loci in each variant string}
\usage{
count_het_loci(x)
}
\arguments{
\item{x}{a variant string or vector of variant strings.}
}
\description{
Count the number of heterozygous loci in each variant string and return as a
vector.
}
