% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{order_position_string}
\alias{order_position_string}
\title{Reorders a position string in alphabetical order of genes}
\usage{
order_position_string(x)
}
\arguments{
\item{x}{a character string or vector of character strings.}
}
\description{
Reorders a position string in alphabetical order of genes. This can be useful
when checking for duplicated strings, as it allows us to account for the same
genes being present in a different order.
}
