% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{position_from_variant_string}
\alias{position_from_variant_string}
\title{Extract a position string from a variant string}
\usage{
position_from_variant_string(x)
}
\arguments{
\item{x}{a character string or vector of character strings.}
}
\description{
Extract a position string from a variant string by stripping the amino acids.
}
